bug fix - isotope items jumping on filtering
so apparently when you use transform on isotope it glitches and the item jumps, to prevent this, I created a new clickable-inner and wrapped the item that I wanna transform inside this, and didn't use the outer wrapper that isotope uses. It took me another 1 hour to figure this all out.